+++ sports-scores/src/components/footer.tsx
@@ -6,7 +6,7 @@ import aflBall from "@/../public/football.svg"
 
 export default function Footer() {
     return (
-    <footer className="bg-gray-200 flex flex-row fixed bottom-0 h-[8vh] w-full p-1 gap-2 overflow-auto border-t">
+    <footer className="bg-gray-200 flex flex-row bottom-0 h-[8%] w-full p-1 gap-2 overflow-auto border-t">
         <Link href="/sports/cricket">
             <div className="size-11 rounded-full bg-gray-500 flex place-content-center">
                 <Image src={cricketBall} width={30} height={30} alt="cricket ball" className="text-center"></Image>
