+++ sports-scores/src/components/header.tsx
@@ -9,7 +9,7 @@ import menu from "@/../public/three-horizontal-lines-icon.svg"
 export default function Header() {
     const router = useRouter()
 
-    return (<header className="bg-gray-200 h-[8vh] fixed w-full flex justify-between border-b">
+    return (<header className="bg-gray-200 w-full h-[8%] flex justify-between border-b">
         <button className="ps-2" onClick={() => router.back()}>
             <Image src={backArrow} width={30} height={30} alt="back arrow"></Image>
         </button>
